Develop

+++ src/main/java/bigsanghyuk/four_uni/comment/service/CommentService.java
@@ -1,12 +1,15 @@
 package bigsanghyuk.four_uni.comment.service;
 
+import bigsanghyuk.four_uni.comment.domain.DeleteCommentInfo;
 import bigsanghyuk.four_uni.comment.domain.EditCommentInfo;
 import bigsanghyuk.four_uni.comment.domain.RegisterCommentInfo;
 import bigsanghyuk.four_uni.comment.domain.entity.Comment;
+import bigsanghyuk.four_uni.comment.domain.entity.LikeComment;
 import bigsanghyuk.four_uni.comment.repository.CommentRepository;
 import bigsanghyuk.four_uni.comment.repository.LikeCommentRepository;
 import bigsanghyuk.four_uni.exception.comment.CommentEditOtherUserException;
 import bigsanghyuk.four_uni.exception.comment.CommentNotFoundException;
+import bigsanghyuk.four_uni.exception.comment.CommentRemoveOtherUserException;
 import bigsanghyuk.four_uni.exception.post.PostNotFoundException;
 import bigsanghyuk.four_uni.exception.user.UserNotFoundException;
 import bigsanghyuk.four_uni.post.repository.PostRepository;
@@ -16,6 +19,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
+import java.util.LinkedList;
 import java.util.List;
 
 @Service
@@ -60,9 +64,19 @@ public class CommentService {
     }
 
     @Transactional
-    public void remove(Long commentId) {
-        commentRepository.findById(commentId)
+    public void remove(Long commentId, @Valid DeleteCommentInfo deleteCommentInfo) {
+        postRepository.findById(deleteCommentInfo.getPostId())
+                .orElseThrow(PostNotFoundException::new);
+
+        commentRepository.findByUserIdOrderByIdDesc(deleteCommentInfo.getUserId())
+                .orElseThrow(UserNotFoundException::new);
+
+        Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(CommentNotFoundException::new);
+
+        if (!deleteCommentInfo.getUserId().equals(comment.getUserId())) {
+            throw new CommentRemoveOtherUserException();
+        }
 
         commentRepository.deleteCommentByAndId(commentId);
         likeCommentRepository.deleteLikeCommentByCommentId(commentId);
@@ -80,6 +94,16 @@ public class CommentService {
         }
 
         comments.addAll(childComments);
+        return comments;
+    }
+
+    public List<Comment> getLikedComment(Long userId) {
+        LinkedList<Comment> comments = new LinkedList<>();
+        List<LikeComment> likedComments = likeCommentRepository.findByUserIdOrderByIdDesc(userId);
+        for (LikeComment likeComment : likedComments) {
+            comments.add(commentRepository.findById(likeComment.getCommentId()).orElseThrow(CommentNotFoundException::new));
+        }
+
         return comments;
     }
 }
